requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
google-generativeai
nltk
PyPDF2
sentence-transformers
transformers
torch